refactor: consolidate h2o code and restore shared cache sizing
